Place RabbitMQ configuration in confd
Files places in confd are loaded after 'rabbitmq.conf' is loaded,
allowing the user configuration to take precedence. The ConfigMap has
keys without any numeric prefix because file naming is done in
StatefulSet's pod template volume mounts. This allows flexibility in the
future, shall we decide we rename or change the numeric prefixes.

This commit also adds a condition to decide whether to include the
server ConfigMap as a volume. This happens if RabbitMQ env or advanced
config are set in the RabbitmqCluster Spec. The server conf is partially
projected into a volume that is always specified.